Refactor progress bars out of invocation summary.
