+++ client/src/components/WorkflowInvocationState/WorkflowInvocationOverview.vue
@@ -2,24 +2,19 @@
 import { BAlert, BButton } from "bootstrap-vue";
 import { computed } from "vue";
 
-import { type InvocationJobsSummary, type InvocationStep, type WorkflowInvocationElementView } from "@/api/invocations";
+import { type InvocationJobsSummary, type WorkflowInvocationElementView } from "@/api/invocations";
 import { useWorkflowInstance } from "@/composables/useWorkflowInstance";
 import { getRootFromIndexLink } from "@/onload";
 import { withPrefix } from "@/utils/redirect";
 
-import {
-    errorCount as jobStatesSummaryErrorCount,
-    jobCount as jobStatesSummaryJobCount,
-    numTerminal,
-    okCount as jobStatesSummaryOkCount,
-    runningCount as jobStatesSummaryRunningCount,
-} from "./util";
+import { runningCount as jobStatesSummaryRunningCount } from "./util";
 
 import ExternalLink from "../ExternalLink.vue";
 import HelpText from "../Help/HelpText.vue";
 import InvocationGraph from "../Workflow/Invocation/Graph/InvocationGraph.vue";
+import InvocationJobsProgressBar from "./InvocationJobsProgressBar.vue";
+import InvocationStepsProgressBar from "./InvocationStepsProgressBar.vue";
 import LoadingSpan from "@/components/LoadingSpan.vue";
-import ProgressBar from "@/components/ProgressBar.vue";
 import InvocationMessage from "@/components/WorkflowInvocationState/InvocationMessage.vue";
 
 function getUrl(path: string): string {
@@ -72,30 +67,6 @@ const disabledReportTooltip = computed(() => {
     }
 });
 
-const stepCount = computed<number>(() => {
-    return props.invocation.steps.length || 0;
-});
-
-type StepStateType = { [state: string]: number };
-
-const stepStates = computed<StepStateType>(() => {
-    const stepStates: StepStateType = {};
-    const steps: InvocationStep[] = props.invocation.steps || [];
-    for (const step of steps) {
-        if (!step) {
-            continue;
-        }
-        // the API defined state here allowing null and undefined is odd...
-        const stepState: string = step.state || "unknown";
-        if (!stepStates[stepState]) {
-            stepStates[stepState] = 1;
-        } else {
-            stepStates[stepState] += 1;
-        }
-    }
-    return stepStates;
-});
-
 const invocationPdfLink = computed<string | null>(() => {
     const id = invocationId.value;
     if (id) {
@@ -111,36 +82,8 @@ const uniqueMessages = computed(() => {
     return Array.from(uniqueMessagesSet).map((message) => JSON.parse(message)) as typeof messages;
 });
 
-const stepStatesStr = computed<string>(() => {
-    return `${stepStates.value?.scheduled || 0} of ${stepCount.value} steps successfully scheduled.`;
-});
-
-const okCount = computed<number>(() => {
-    return jobStatesSummaryOkCount(props.jobStatesSummary);
-});
-
 const runningCount = computed<number>(() => {
     return jobStatesSummaryRunningCount(props.jobStatesSummary);
-});
-
-const jobCount = computed<number>(() => {
-    return jobStatesSummaryJobCount(props.jobStatesSummary);
-});
-
-const errorCount = computed<number>(() => {
-    return jobStatesSummaryErrorCount(props.jobStatesSummary);
-});
-
-const newCount = computed<number>(() => {
-    return jobCount.value - okCount.value - runningCount.value - errorCount.value;
-});
-
-const jobStatesStr = computed(() => {
-    let jobStr = `${numTerminal(props.jobStatesSummary) || 0} of ${jobCount.value} jobs complete`;
-    if (!props.invocationSchedulingTerminal) {
-        jobStr += " (total number of jobs will change until all steps fully scheduled)";
-    }
-    return `${jobStr}.`;
 });
 
 const emit = defineEmits<{
@@ -188,11 +131,6 @@ function onCancel() {
                 </BButton>
             </div>
             <div class="w-100">
-                <ProgressBar
-                    v-if="!stepCount"
-                    note="Loading step state summary..."
-                    :loading="true"
-                    class="steps-progress" />
                 <template v-if="uniqueMessages.length">
                     <InvocationMessage
                         v-for="message in uniqueMessages"
@@ -202,32 +140,15 @@ function onCancel() {
                         :invocation="invocation">
                     </InvocationMessage>
                 </template>
-                <ProgressBar
-                    v-else-if="invocationState == 'cancelled'"
-                    note="Invocation scheduling cancelled - expected jobs and outputs may not be generated."
-                    :error-count="1"
-                    class="steps-progress" />
-                <ProgressBar
-                    v-else-if="invocationState == 'failed'"
-                    note="Invocation scheduling failed - Galaxy administrator may have additional details in logs."
-                    :error-count="1"
-                    class="steps-progress" />
-                <ProgressBar
+                <InvocationStepsProgressBar
                     v-else
-                    :note="stepStatesStr"
-                    :total="stepCount"
-                    :ok-count="stepStates.scheduled"
-                    :loading="!invocationSchedulingTerminal"
-                    class="steps-progress" />
-                <ProgressBar
-                    :note="jobStatesStr"
-                    :total="jobCount"
-                    :ok-count="okCount"
-                    :running-count="runningCount"
-                    :new-count="newCount"
-                    :error-count="errorCount"
-                    :loading="!invocationAndJobTerminal"
-                    class="jobs-progress" />
+                    :invocation="invocation"
+                    :invocation-state="invocationState"
+                    :invocation-scheduling-terminal="invocationSchedulingTerminal" />
+                <InvocationJobsProgressBar
+                    :job-states-summary="jobStatesSummary"
+                    :invocation-scheduling-terminal="invocationSchedulingTerminal"
+                    :invocation-and-job-terminal="invocationAndJobTerminal" />
             </div>
         </div>
         <!-- Once the workflow for the invocation has been loaded, display the graph -->
